Refactor TranscriptWriter into small components.
