+++ src/views/components/TranscriptWriter/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './TranscriptWriter'
+export * from './TranscriptWriter'
